Some fixes for disco to work and special chars
There was a comment about needing JSON which was
indeed a bug. Trivial fix.

Also Discovery can throw an exception on some input
(found that with a carat) so try/except that.

Also hit a character encoding error in a debug
print so fixed that.

Took out a little of the commented out code now
that the Discovery integration is in. There is
still some test code and fake discovery but I am
leaving that in case it is still wanted.