Limit the discovery response
Added a count limit on the query. Hard-coded to 2 so top 2 will
be in the log.

Have the formatter limit set to 1 so we only get 1 result.

Use 'text' instead of the blekko snippet and title. Might
look better. We'll see.

Truncate the text. Hard-coded at 500. You'll see ... if truncation
happened.