Make WOS deployable with flask server
Adds a python flask server in a subdir.
The slack integration is started if the necessary
config exists. The web-ui is able to run and
report an error when env vars are missing.

Some refactoring to share code between the flask
server (server.py) and the command-line run.py.

Use PORT env var for Bluemix health check
Bluemix expects the server on the PORT.
So use the env var PORT when starting it.