Make WOS deployable with Bluemix
Add services in manifest. Get env vars from VCAP. Get BOT_ID from user.

Allow SLACK_BOT_USER so that set-up doesn't need the
script to get an ID from a user name (do it in the code).

Don't call init_customer over and over with nothing.

Adds a python flask server in a subdir.
The slack integration is started if the necessary
config exists. The web-ui is able to run and
report an error when env vars are missing.

Some refactoring to share code between the flask
server (server.py) and the command-line run.py.

Use PORT env var for Bluemix health check
Bluemix expects the server on the PORT.
So use the env var PORT when starting it.

Use separate instances for Web UI and Slack.