Add web UI to watson-online-store
Adds a python flask server in a subdir.
The slack integration is started if the necessary
config exists. The web-ui is able to run and
report an error when env vars are missing.

Refactor user profile so Slack/Web-UI can impl.
Web UI uses bogus user info.

Also refactor conversation loop so web UI and Slack
share more code.

Unfurl images for IBM store
Split out the URLs and display them as images. This is
pretty specific to the IBM store URLs for now.