Perform login with Slack user, create user if needed
Using the authenticated Slack user, get the Slack user
email address and check the Cloudant DB for that user.
If user exists, log them in and get their Customer info.
If user does not exist, create one.
Context will be set using this Customer info.
Code is removed to handle login and adding a user.