Add new cust to DB in init_customer(). Don't init with empty.
The init_customer() case where a new customer is created was
not adding the customer to the DB. Add'em.
The init_customer() was also called often in our run loop
without a user_id. Only call when we have one.
Also refactored some of the redundant conditions catching
a few more of the error cases.
Added unit tests.