For direct message, don't require the at-bot-id. Just chat.
First check for at-botid so that we can remove that from the text.
Also fix so that the at-botid doesn't have to be first. It can
be tacked on at the end or wherever to get the bot's attention.
If there are multiple mentions remove them all.
There might be a little extra whitespace, but no harm.

If there is not an at-botid, then check to see if it is
a direct message. A channel that starts with a D is a
direct message. For direct message you don't need to
type that at-botid.