Add images to results.
This finds and adds the URL to the image.
It's shrunk down pretty small on purpose. There's
a regex replace that does that if you want to resize.
Added slack_encode to catch < & > chars that would break
things, but that wasn't really the problem here.

Output to Slack will use the image_url, as will output to Watson
for Discovery results. Output to Watson for List Shopping Cart
will use the product URL, and this URL is Added to (and used for
Delete from) the Shopping Cart.