Verify conversation workspace exists. Create/upload if needed.
If WORKSPACE_ID is specified, verify that the workspace exists
else throw an Exception (abort). This is typically how devs
run today.

Else if WORKSPACE_ID is not specified, then look for it by name.
Default name is 'watson-online-store', but allow env var
WORKSPACE_NAME to override it (useful for dev/test).
The lookup by name is needed mostly because if we create one, we need
to know how to find it and will not have a configured WORKSPACE_ID.

When we don't have a WORKSPACE_ID and can't find one by name (e.g.,
one we already created) then create one using the data/workspace.json.

Note:  The intended path is that a Bluemix wos-conversation-service
is created when this is deployed. So it will be empty and we will
create the workspace. On restarts, we'll find the one that was
created.