Mention that "View app" doesn't work in the README.
During deployment and if you navigate back to the toolchain
there is a "View app" button. Sometimes it shows a proper
message about "No UI..." but it currently it can get stuck
with an empty message (and does nothing). So add a sentence
to tell folks not to touch that.

Also deploy now has a "Create" button so mention they. You
do have to touch that one if using the Bluemix button.