Fix Issue with bad input.
We were getting a stacktrace due to bad input containing newlines, which
causes a Traceback.
One issue fixed is that we no longer need to make a "dummy" call to initialize
the Watson Assistant context. Doing this moved us forward in the Assistant
state machine, causing errors in how we handled the context.
Another issue is that the app was considering text from the bot in Slack
(which came from Watson Assistant) as user input. This also moved us forward
in the dialog, and resulted in sending the newlines that caused the crash.

Closes: #176